undim {listenv}	R Documentation

Removes the dimension of an object

Description

Removes the dimension of an object

Usage

undim(x, ...)

Arguments

x
An object with or without dimensions

...
Not used.

Details

This function does attr(x, "dim") <- NULL, which automatically also does attr(x, "dimnames") <- NULL. However, other attributes such as names attributes are preserved, which is not the case if one does dim(x) <- NULL.

Value

The object with the dimension attribute removed.

[Package listenv version 0.10.1 Index]